ass/requirements.txt
fastapi[all]>=0.100.0
PyMuPDF>=1.23.0
pytesseract>=0.3.10
Pillow>=10.2.0
python-docx>=1.1.0
pandas>=2.0.0
uvicorn>=0.23.0
python-multipart>=0.0.9
ultralytics>=8.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
openai>=1.0.0
python-dotenv>=1.0.0
